Fix existing file removal

+++ R/mt_download.r
@@ -5,23 +5,17 @@
 #'
 #' @param download_link Download Link
 #' @param file_path File Path
+#' @param overwrite If to overwrite the existing file
 #'
 #' @export
 
 mt_download <- function(download_link, file_path, overwrite = F) {
-  if (endsWith(file_path, ".csv")) {
-    gz_file_path <- paste0(file_path, ".gz")
-  } else if (endsWith(file_path, ".gz")) {
-    gz_file_path <- file_path
-  } else {
-    stop("Incorrect file_path argument - .csv or .gz file format required")
-  }
+  gz_file_path <- paste0(file_path, ".gz")
 
-  download.file(download_link, gz_file_path)
-
-  if (isTRUE(overwrite)) {
+  if (file.exists(file_path) && isTRUE(overwrite)) {
     file.remove(file_path)
   }
 
+  download.file(download_link, gz_file_path)
   gunzip(gz_file_path)
 }
